admin/수정

+++ backend/src/main/java/com/example/backend/dto/admin/AdminProductDetailDto.java
@@ -2,6 +2,7 @@ package com.example.backend.dto.admin;
 
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -16,7 +17,7 @@ public class AdminProductDetailDto {
     private String productBrand;
     private String productName;
     private String modelNum;
-    private Long originalPrice;
+    private BigDecimal originalPrice;
     private int productQuantity;
     private String productSize;
